backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
ultralytics==8.3.0
python-multipart==0.0.20
pillow==11.0.0
pydantic==2.10.0
websockets==14.1
python-dotenv==1.0.0
google-generativeai==0.8.0
httpx==0.27.0
paho-mqtt==1.6.1
pyserial==3.5
